pyproject.toml
[tool.poetry]
name = "webqa-databuilder"
version = "0.1.0"
description = ""
authors = ["zjrwtx <3038880699@qq.com>"]
license = "GPL"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.25.0"
python-dotenv = "^1.0.1"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
